registry/requirements.txt
Flask
mysql-connector-python
PyJWT
requests
gunicorn
bcrypt
dotenv
flask_apscheduler